Small changes to player movement, not yet applied
See the change in the movement test scene. I was thinking maybe we
should putoff the change until after the playtest and figure out the
best way to change player movement. I don't want the level to be
redesigned to suit the new movement change. That would be too much work
for Monday.